+++ src/scripts/userscript.ts
@@ -1,6 +1,9 @@
 import type { PxlsApp } from '../pxls/pxls-global';
 
 export abstract class PxlsUserscript {
+    readonly requiresVirginmap: boolean = false;
+    readonly requiresHeatmap: boolean = false;
+
     protected constructor(
         readonly name: string,
         readonly beforeApp?: () => void,